Animations for the crafting ritual
they are under fuel, with (1) variations of Brown Wood and Stone holding the switching ones animations, with (1) of Herb switching both herbs and the flower, and (2) of Herb switch just the two herbs. IDK where to put them so they are just in there for now.

wrapped in juaning wind
hearts will never bri
frozen and banished
out of siddi